requirements.txt
fastapi
uvicorn
anthropic
langgraph
langchain-anthropic
pydantic
python-dotenv
supabase
semgrep
bandit
ruff
json-repair